Adding Dask LSF Job Queue Backend for better HPC Integration

+++ propertyestimator/utils/packmol.py
@@ -7,15 +7,15 @@ Based on the `SolvationToolkit <https://github.com/MobleyLab/SolvationToolkit>`_
 """
 
 import logging
-import mdtraj
 import os
 import random
+import shutil
 import string
 import subprocess
-import shutil
-
 from distutils.spawn import find_executable
+from tempfile import mkdtemp
 
+import mdtraj
 from simtk import openmm
 from simtk import unit
 from simtk.openmm import app
@@ -44,7 +44,9 @@ def pack_box(molecules,
              tolerance=2.0,
              box_size=None,
              mass_density=None,
-             verbose=False):
+             verbose=False,
+             working_directory=None,
+             retain_working_files=False):
 
     """Run packmol to generate a box containing a mixture of molecules.
 
@@ -65,6 +67,12 @@ def pack_box(molecules,
         Target mass density for final system, if available.
     verbose : bool, optional, default=False
         If True, verbose output is written.
+    working_directory: str, optional
+        The directory in which to generate the temporary working files. If none,
+        a temporary one will be created.
+    retain_working_files: bool
+        If True, all of the working files, such as individual molecule coordinate
+        files will be retained.
 
     Returns
     -------
@@ -80,6 +88,13 @@ def pack_box(molecules,
     if len(molecules) != len(n_copies):
         raise ValueError("Length of 'molecules' and 'n_copies' must be identical")
 
+    temporary_directory = False
+
+    if working_directory is None:
+
+        working_directory = mkdtemp()
+        temporary_directory = True
+
     # Create PDB files for all components
     pdb_filenames = list()
     pdb_flavor = oechem.OEOFlavor_PDB_Default
@@ -88,7 +103,7 @@ def pack_box(molecules,
 
     for index, molecule in enumerate(molecules):
 
-        tmp_filename = "{}.pdb".format(index)
+        tmp_filename = os.path.join(working_directory, f'{index}.pdb')
         pdb_filenames.append(tmp_filename)
 
         # Write PDB file
@@ -108,22 +123,13 @@ def pack_box(molecules,
             file.write(pdb_contents.encode())
 
         oe_pdb = mdtraj.load_pdb(tmp_filename)
-
-        # We definitely want to make sure the temp pdb file is identical to
-        # that stored in the topology to ensure the correct generation
-        # of CONNECT statements.
-        tmp_mdtraj_filename = "tmp_mdtraj.pdb"
-        oe_pdb.save_pdb(tmp_mdtraj_filename)
-
-        os.unlink(tmp_mdtraj_filename)
-
         mdtraj_topologies.append(oe_pdb.topology)
 
     # Run packmol
     if PACKMOL_PATH is None:
         raise IOError("Packmol not found, cannot run pack_box()")
 
-    output_filename = "output.pdb"
+    output_filename = os.path.join(working_directory, "packmol_output.pdb")
 
     # Approximate volume to initialize box
     if box_size is None:
@@ -148,10 +154,10 @@ def pack_box(molecules,
                                        unitless_box_angstrom)
 
     if verbose:
-        print(header)
+        logging.info(header)
 
     # Write packmol input
-    packmol_filename = "packmol_input.txt"
+    packmol_filename = os.path.join(working_directory, "packmol_input.txt")
 
     with open(packmol_filename, 'w') as file_handle:
         file_handle.write(header)
@@ -165,19 +171,27 @@ def pack_box(molecules,
                                          stderr=subprocess.STDOUT).decode("utf-8")
 
         if verbose:
-            print(result)
+            logging.info(result)
 
         packmol_succeeded = result.find('Success!') > 0
 
-    os.unlink(packmol_filename)
+    if not retain_working_files:
 
-    for filename in pdb_filenames:
-        os.unlink(filename)
+        os.unlink(packmol_filename)
+
+        for filename in pdb_filenames:
+            os.unlink(filename)
 
     if not packmol_succeeded:
 
-        logging.warning("Packmol failed to converge")
-        os.unlink(output_filename)
+        if verbose:
+            logging.info("Packmol failed to converge")
+
+        if os.path.isfile(output_filename):
+            os.unlink(output_filename)
+
+        if not retain_working_files and temporary_directory:
+            shutil.rmtree(working_directory)
 
         return None, None
 
@@ -187,15 +201,18 @@ def pack_box(molecules,
 
     # Read the resulting PDB file.
     pdbfile = app.PDBFile(output_filename)
-    os.unlink(output_filename)
+
+    if not retain_working_files:
+        os.unlink(output_filename)
+
+        if temporary_directory:
+            shutil.rmtree(working_directory)
 
     # Extract topology and positions
     topology = pdbfile.getTopology()
     positions = pdbfile.getPositions()
 
     unitless_box_nm = box_size / unit.nanometers
-    # import numpy as np
-    # box_vectors = np.diag([box_size]*3)
 
     box_vector_x = openmm.Vec3(unitless_box_nm, 0, 0)
     box_vector_y = openmm.Vec3(0, unitless_box_nm, 0)
@@ -203,7 +220,6 @@ def pack_box(molecules,
 
     # Set the periodic box vectors.
     topology.setPeriodicBoxVectors([box_vector_x, box_vector_y, box_vector_z] * unit.nanometers)
-    # topology.setPeriodicBoxVectors(box_vectors)
 
     return topology, positions
